[refactor] rename AddLectureSWTab to AddOrUpdateLectureSWTab
- Because AddLectureSWTab includes modify and delete feature, rename component name to AddOrUpdateLectureSWTab
